+++ spec/metanorma/biblio_spec.rb
@@ -905,14 +905,14 @@ RSpec.describe Metanorma::Standoc do
            <references id="_" normative="true" obligation="informative">
              <title>Normative references</title>
              <p id="_">The following documents are referred to in the text in such a way that some or all of their content constitutes requirements of this document. For dated references, only the edition cited applies. For undated references, the latest edition of the referenced document (including any amendments) applies.</p>
-             <bibitem id="A" type="inbook">
-               <formattedref format="application/x-isodoc+xml">Wozniak, S., Jobs, Steve &amp; Hoover, J. Edgar. 1991-1992. 1996-01-02. <em>Work</em>. Gates, W. H &amp; UNICEF, Collected Essays. <em>Bibliographers Anonymous</em>. ISO 1234. Geneva: International Standardization Organization. <link target="http://www.example.com"/>. 4 2–3 12-13 19. </formattedref>
+              <bibitem id="A" type="inbook">
+               <formattedref format="application/x-isodoc+xml">Wozniak, S., Jobs, Steve &amp; Hoover, J. Edgar. 1991-1992. 1996-01-02. <em>Work</em>. Gates, W. H &amp; UNICEF, Collected Essays. <em>Bibliographers Anonymous</em>. 4, draft. Also available in paperback. ISO 1234. Geneva: International Standardization Organization. <link target="http://www.example.com"/>. 4 2–3 12-13 19. </formattedref>
                <title>
                  <em>Work</em>
                </title>
                <uri type="citation">http://www.example.com</uri>
-               <docidentifier>B</docidentifier>
                <docidentifier type="ISO">ISO 1234</docidentifier>
+               <docidentifier>B</docidentifier>
                <date type="issued">
                  <from>1991</from>
                  <to>1992</to>
@@ -954,6 +954,9 @@ RSpec.describe Metanorma::Standoc do
                    <name>International Standardization Organization</name>
                  </organization>
                </contributor>
+               <edition>4</edition>
+               <version>draft</version>
+               <note>Also available in paperback.</note>
                <place>Geneva</place>
                <relation type="includedIn">
                  <bibitem type="book">
